+++ src/main/java/backend/wal/auth/application/port/in/IssueTokenUseCase.java
@@ -1,7 +1,5 @@
 package backend.wal.auth.application.port.in;
 
-import backend.wal.auth.application.port.out.TokenResponseDto;
-
 public interface IssueTokenUseCase {
 
     TokenResponseDto issue(Long userId);
